GUI: improve dialog text visibility and style
Dialog text in the GUI was difficult to read so changed the user text to white, Sakura text to light pink,
and set both to bold 16px Arial. Changed the plain background to a more aesthetic one.

These changes improve the user experience by making messages
clearer and easier to read.